This change is an attempt at reworking the storage layer to use a single large memory-mapped file in place of per-series files.  It's very much a work in progress, and breaks many things.  Known issues:
1. No support yet for expiring data, and if attempted probably terrible things
will happen.

2. No support yet for syncing data, syncStragegy is always Never.

3. It's not possible to have more than 2GB worth of samples.  This would be easy
to fix for the non-32bit case, but in order to accomodate it I plan to add support
for multiple chunk mmaps.

4. Only doubleDeltaEncodedChunk uses mmap backing so far.  It should be trivial to
adapt the others, I just haven't got around to it.

5. Many tests are broken or disabled.

6. There may be a slight performance degradation.  My tests against changeset
a9b96be3fdb49593c47bc827e43f538990aa6ddc (with or without my patch) using
"prombench -scrape-interval 50ms -test-duration 5m -exporters oscillate:1 --
-storage.local.retention 15m" yielded 200K samples/s.  After updating to
4520e124403b234ef973636dc4c17e197ad423b4, stock prometheus yields 175K, and
with my patch 165K.

The general idea is to allocate blocks ("extents") of N chunks at a time whenever
persisting series data, and to use an index to map fingerprints to extent lists.
When it comes to implementing expiry, we'll recycle fully empty extents and add
a new index to track them.

The chunk reading code is mostly the same except it now reads from the mmap and
the resulting chunks are backed by it, i.e. their memory doesn't come from the
heap.

The chunk writing code is also largely unchanged, except now it writes to an
offset in the big file based on the extents assigned in the index.  Upon
success it reads the chunks back from the mmap so that they're not on the heap.

Dropped all chunk eviction and pinning code (chunkDesc eviction lives on for
now.)  chunkDesc.isEvicted() still exists but it no longer locks and it will
never return true once it's returned false.  In other words it's used to
identify chunkDescs for which we've never loaded the chunk.

New indexing queue for managing fp-to-extent index.

New storagetool option "dump-chunks", to read a series or chunks file.