First pass at Issue #2174: Add support for local storage engine that only supports federation queries.
Adds a new option for flag -storage.local.engine: "federated".  When federated
mode is selected, turns on new local storage option DoNotPersistChunks.

This changeset bumps the checkpoint format to 3 in order to add a new
per-series field lastSampleValue, since checkpoints don't include any chunks in
federated mode.  Checkpoint formats 1 and 2 are still supported for reads.
This is the only change that applies both to the "federated" and "persisted"
settings for -storage.local.engine.

All of the following behaviour applies only if DoNotPersistChunks=true:

* Non-federation queries don't work.

* No chunk eviction or throttling (rushed mode).

* No chunks are ever encoded or persisted.  When adding a sample to a series
  we do nothing except update lastTime, savedFirstTime, lastSampleValue,
  lastSampleValueSet.

* During series maintenance, skip all the chunk related stuff, just drop any
  series with lastTime is too old.  Never archive, unarchive, or quarantine
  series if DoNotPersistChunks=true.

* Crash recovery ignores series files.